+++ testlib/src/main/java/com/diffplug/spotless/tag/BlackTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 DiffPlug
+ * Copyright 2021-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,5 +26,5 @@ import org.junit.jupiter.api.Tag;
 
 @Target({TYPE, METHOD})
 @Retention(RUNTIME)
-@Tag("Black")
+@Tag("black")
 public @interface BlackTest {}