+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/GitRatchetGradleTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2023 DiffPlug
+ * Copyright 2016-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -36,6 +36,9 @@ import org.gradle.testkit.runner.TaskOutcome;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import com.diffplug.spotless.ClearGitConfig;
+
+@ClearGitConfig
 class GitRatchetGradleTest extends GradleIntegrationHarness {
 	private static final String TEST_PATH = "src/markdown/test.md";
 
@@ -268,10 +271,9 @@ class GitRatchetGradleTest extends GradleIntegrationHarness {
 	private RevCommit addAndCommit(Git git) throws NoFilepatternException, GitAPIException {
 		PersonIdent emptyPerson = new PersonIdent("jane doe", "jane@doe.com", new Date(0), TimeZone.getTimeZone("UTC"));
 		git.add().addFilepattern(".").call();
-		RevCommit commit = git.commit().setMessage("baseline")
+		return git.commit().setMessage("baseline")
 				.setCommitter(emptyPerson)
 				.setAuthor(emptyPerson)
 				.call();
-		return commit;
 	}
 }